added documentation, added JSON-indentation
