+++ .gitignore
@@ -3,7 +3,6 @@
 ##
 ## Get latest from https://github.com/github/gitignore/blob/master/VisualStudio.gitignore
 
-airports.csv
 GraphicProto.txt
 
 # User-specific files
