+++ Api/Controllers/Airac/GetCurrent.cs
@@ -12,12 +12,12 @@
             {
                 if (airacs[i + 1].StartDate > dateNow)
                 {
-                    return new JsonResult(airacs[i]);
+                    return new JsonResult(airacs[i], Options.JsonOptions);
                 }
             }
 
             //If you are here then you are somehow later in time than the publication of any airac (or you have time travelled)
-            return new JsonResult(new ApiError("You are either a time traveller or there are no published AIRAC Cycles"));
+            return new JsonResult(new ApiError("You are either a time traveller or there are no published AIRAC Cycles"), Options.JsonOptions);
         }
     }
 }
