+++ Api/Controllers/Airac/GetNext.cs
@@ -10,7 +10,7 @@
 
             if (dateNow > airacs.Last().StartDate)
             {
-                return new JsonResult(new ApiError("The next Airac is not available yet"));
+                return new JsonResult(new ApiError("The next Airac is not available yet"), Options.JsonOptions);
             }
 
             for (var i = 0; i < airacs.Count; i++)
@@ -22,7 +22,7 @@
             }
 
             //If you are here then you are somehow later in time than the publication of any airac (or you have time travelled)
-            return new JsonResult(new ApiError("You are either a time traveller or there are no published AIRAC Cycles"));
+            return new JsonResult(new ApiError("You are either a time traveller or there are no published AIRAC Cycles"), Options.JsonOptions);
         }
     }
 }
