+++ Api/Controllers/Airlines/GetSearchesMultipleParameters.cs
@@ -13,13 +13,17 @@ namespace Api.Controllers.Airlines
                 return new JsonResult(
                     new ApiError(
                         "Please use one of the following search Parameters: name=, iata=, icao=, callsign=, country=, active="
-                    )
+                    ),
+                    Options.JsonOptions
                 );
             }
 
             if (matches.Count < 2)
             {
-                return new JsonResult(new ApiError("Please specify at lease 2 search paramters"));
+                return new JsonResult(
+                    new ApiError("Please specify at lease 2 search paramters"),
+                    Options.JsonOptions
+                );
             }
 
             var airlines = AirlinesJson.ReadJson();
@@ -27,8 +31,14 @@ namespace Api.Controllers.Airlines
             var searches = Searches.GetSearchResults(matches);
 
             return matchAll
-              ? new JsonResult(SearchResults.Get(searches, airlines, Parameters.MatchAll))
-              : new JsonResult(SearchResults.Get(searches, airlines, Parameters.MatchAny));
+              ? new JsonResult(
+                    SearchResults.Get(searches, airlines, Parameters.MatchAll),
+                    Options.JsonOptions
+                )
+              : new JsonResult(
+                    SearchResults.Get(searches, airlines, Parameters.MatchAny),
+                    Options.JsonOptions
+                );
         }
     }
 }
