+++ Api/Controllers/Hoppie/HoppieController.cs
@@ -4,14 +4,29 @@
     [ApiController]
     public class HoppieController : Controller
     {
+        /// <summary>
+        /// Get all current online stations on the Hoppie-Network
+        /// </summary>
+        /// <returns></returns>
         [HttpGet("/hoppie/stations")]
         [ResponseCache(VaryByHeader = "User-Agent", Duration = 30)]
-        public JsonResult GetAllStations() => Json(AllClients.Get());
+        public JsonResult GetAllStations() => Json(AllClients.Get(), Options.JsonOptions);
 
+        /// <summary>
+        /// Get a specific Hoppie-Station by Callsign
+        /// </summary>
+        /// <param name="station"></param>
+        /// <returns></returns>
         [HttpGet("/hoppie/stations/{station}")]
         [ResponseCache(VaryByHeader = "User-Agent", Duration = 30)]
-        public JsonResult GetStation(string station = "") => Json(SingleStation.Get(station));
+        public JsonResult GetStation(string station = "") => Json(SingleStation.Get(station), Options.JsonOptions);
 
+        /// <summary>
+        /// Get the conversation of either a single station with other stations or between two specific stations
+        /// </summary>
+        /// <param name="stationOne"></param>
+        /// <param name="stationTwo"></param>
+        /// <returns></returns>
         [HttpGet("/hoppie/conversation/{stationOne}/{stationTwo?}")]
         [ResponseCache(VaryByHeader = "User-Agent", Duration = 30)]
         public string GetConversation(string stationOne = "", string? stationTwo = null) => Conversation.Get(stationOne, stationTwo);
