+++ Api/Controllers/Vatsim/Stats/StatsController.cs
@@ -6,6 +6,11 @@ namespace Api.Controllers.Vatsim.Stats
     [ApiController]
     public class StatsController : Controller
     {
+        /// <summary>
+        /// Get the Vatsim-Stats for a CID
+        /// </summary>
+        /// <param name="cid"></param>
+        /// <returns></returns>
         [HttpGet("/vatsim/stats/{cid}")]
         [ResponseCache(VaryByHeader = "User-Agent", Duration = 30)]
         public async Task<JsonResult> Get(string cid)
@@ -14,11 +19,10 @@ namespace Api.Controllers.Vatsim.Stats
 
             if (!isCid || !VatsimData.DoesCIDExist(cidParsed))
             {
-                return Json("Input was not a valid CID");
+                return Json(new ApiError("Input was not a valid CID"));
             }
 
-            var client = new HttpClient();
-            var stats = await client.GetFromJsonAsync<VatsimStats>($"https://api.vatsim.net/api/ratings/{cidParsed}/");
+            var stats = AviationSharp.Vatsim.Stats.VatsimStats.GetFullStats(cidParsed);
 
             return Json(stats, Options.JsonOptions);
         }
